+++ changelog.md
@@ -105,3 +105,4 @@
 
 * Fixed issue with natural breeding
 * Blueprint will now better display vertically stacked buildings
+* Villager preview in the editor is now animated